feat: remove old modal
