+++ src/connect/components/drive-settings-modal/drive-settings-modal.tsx
@@ -1,4 +1,4 @@
-import { Modal } from '@/powerhouse/components/new-modal';
+import { Modal } from '@/powerhouse';
 import { ComponentPropsWithoutRef } from 'react';
 import { twMerge } from 'tailwind-merge';
 import { Divider } from '..';
