+++ src/powerhouse/components/modal/modal.stories.tsx
@@ -1,48 +1,162 @@
-import type { Meta, StoryObj } from '@storybook/react';
-import { Button } from 'react-aria-components';
-import { Modal } from './modal';
+import { DropdownMenu } from '@/powerhouse';
+import { useArgs } from '@storybook/preview-api';
+import { Meta, StoryObj } from '@storybook/react';
+import { Modal } from '.';
 
-const meta: Meta<typeof Modal> = {
-    title: 'Powerhouse/Components/Modal',
+const meta = {
+    title: 'Powerhouse/Components/NewModal',
     component: Modal,
-    argTypes: {
-        open: { control: { type: 'boolean' } },
-        children: { control: { type: 'text' } },
-        onClose: { control: { type: 'action' } },
-        dialogProps: { control: { type: 'object' } },
-        modalProps: { control: { type: 'object' } },
+    parameters: {
+        layout: 'fullscreen',
     },
-};
+} satisfies Meta<typeof Modal>;
 
 export default meta;
+
 type Story = StoryObj<typeof meta>;
 
-const children = (
-    <div className="flex h-[300px] w-[465px] flex-col justify-between p-10">
-        <div className="text-xl">Header</div>
-        <div>
-            Lorem ipsum dolor sit amet consectetur adipisicing elit. Labore, ea.
-            Quo deserunt amet iusto iste? Magni, a distinctio? Aperiam, sapiente
-            ipsam. Sapiente corrupti explicabo deserunt nisi sit cupiditate
-            maxime quis.
-        </div>
-        <div>
-            <Button className="rounded bg-red-400 p-2 text-white">
-                Cancel
-            </Button>
-        </div>
-    </div>
-);
-
-export const Primary: Story = {
+export const Default: Story = {
     args: {
         open: true,
-        children,
-        modalProps: {
-            className: 'top-10',
-        },
-        dialogProps: {
-            className: 'rounded-3xl',
-        },
+    },
+    render: function Wrapper(args) {
+        const [, setArgs] = useArgs<typeof args>();
+
+        return (
+            <div className="h-screen w-screen bg-white">
+                <Modal
+                    {...args}
+                    open={args.open}
+                    onOpenChange={open => {
+                        setArgs({ ...args, open });
+                    }}
+                >
+                    <div>
+                        <DropdownMenu
+                            onItemClick={id => alert(id)}
+                            className="mb-3 block items-center justify-center rounded bg-blue-500 text-white"
+                            menuClassName="border-2 border-indigo-600 w-64 rounded bg-white cursor-pointer"
+                            menuItemClassName="hover:bg-gray-200 px-2"
+                            items={[
+                                {
+                                    id: 'item-1',
+                                    content: 'Item 1',
+                                },
+                                {
+                                    id: 'item-2',
+                                    content: 'Item 2',
+                                },
+                                {
+                                    id: 'item-3',
+                                    content: 'Item 3',
+                                },
+                            ]}
+                        >
+                            open menu
+                        </DropdownMenu>
+                        lorem ipsum dolor sit amet consectetur adipisicing elit.
+                        Quisquam quod, voluptatum, quae voluptatem voluptas
+                        lorem ipsum dolor sit amet consectetur adipisicing elit.
+                        Quisquam quod, voluptatum, quae voluptatem voluptas
+                        lorem ipsum dolor sit amet consectetur adipisicing elit.
+                        Quisquam quod, voluptatum, quae voluptatem voluptaslorem
+                        ipsum dolor sit amet consectetur adipisicing elit.
+                        Quisquam quod, voluptatum, quae voluptatem voluptaslorem
+                        ipsum dolor sit amet consectetur adipisicing elit.
+                        Quisquam quod, voluptatum, quae voluptatem voluptaslorem
+                        ipsum dolor sit amet consectetur adipisicing elit.
+                        Quisquam quod, voluptatum, quae voluptatem voluptaslorem
+                        ipsum dolor sit amet consectetur adipisicing elit.
+                        Quisquam quod, voluptatum, quae voluptatem voluptaslorem
+                        ipsum dolor sit amet consectetur adipisicing elit.
+                        Quisquam quod, voluptatum, quae voluptatem voluptaslorem
+                        ipsum dolor sit amet consectetur adipisicing elit.
+                        Quisquam quod, voluptatum, quae voluptatem
+                        voluptasloremquae voluptatem voluptaslorem ipsum dolor
+                        sit amet consectetur adipisicing elit. Quisquam quod,
+                        voluptatum, quae voluptatem voluptaslorem ipsum dolor
+                        sit amet consectetur adipisicing elit. Quisquam quod,
+                        voluptatum, quae voluptatem voluptasloremquae voluptatem
+                        voluptaslorem ipsum dolor sit amet consectetur
+                        adipisicing elit. Quisquam quod, voluptatum, quae
+                        voluptatem voluptaslorem ipsum dolor sit amet
+                        consectetur adipisicing elit. Quisquam quod, voluptatum,
+                        quae voluptatem voluptasloremquae voluptatem
+                        voluptaslorem ipsum dolor sit amet consectetur
+                        adipisicing elit. Quisquam quod, voluptatum, quae
+                        voluptatem voluptaslorem ipsum dolor sit amet
+                        consectetur adipisicing elit. Quisquam quod, voluptatum,
+                        quae voluptatem voluptasloremquae voluptatem
+                        voluptaslorem ipsum dolor sit amet consectetur
+                        adipisicing elit. Quisquam quod, voluptatum, quae
+                        voluptatem voluptaslorem ipsum dolor sit amet
+                        consectetur adipisicing elit. Quisquam quod, voluptatum,
+                        quae voluptatem voluptasloremquae voluptatem
+                        voluptaslorem ipsum dolor sit amet consectetur
+                        adipisicing elit. Quisquam quod, voluptatum, quae
+                        voluptatem voluptaslorem ipsum dolor sit amet
+                        consectetur adipisicing elit. Quisquam quod, voluptatum,
+                        quae voluptatem voluptasloremquae voluptatem
+                        voluptaslorem ipsum dolor sit amet consectetur
+                        adipisicing elit. Quisquam quod, voluptatum, quae
+                        voluptatem voluptaslorem ipsum dolor sit amet
+                        consectetur adipisicing elit. Quisquam quod, voluptatum,
+                        quae voluptatem voluptasloremquae voluptatem
+                        voluptaslorem ipsum dolor sit amet consectetur
+                        adipisicing elit. Quisquam quod, voluptatum, quae
+                        voluptatem voluptaslorem ipsum dolor sit amet
+                        consectetur adipisicing elit. Quisquam quod, voluptatum,
+                        quae voluptatem voluptasloremquae voluptatem
+                        voluptaslorem ipsum dolor sit amet consectetur
+                        adipisicing elit. Quisquam quod, voluptatum, quae
+                        voluptatem voluptaslorem ipsum dolor sit amet
+                        consectetur adipisicing elit. Quisquam quod, voluptatum,
+                        quae voluptatem voluptasloremquae voluptatem
+                        voluptaslorem ipsum dolor sit amet consectetur
+                        adipisicing elit. Quisquam quod, voluptatum, quae
+                        voluptatem voluptaslorem ipsum dolor sit amet
+                        consectetur adipisicing elit. Quisquam quod, voluptatum,
+                        quae voluptatem voluptasloremquae voluptatem
+                        voluptaslorem ipsum dolor sit amet consectetur
+                        adipisicing elit. Quisquam quod, voluptatum, quae
+                        voluptatem voluptaslorem ipsum dolor sit amet
+                        consectetur adipisicing elit. Quisquam quod, voluptatum,
+                        quae voluptatem voluptasloremquae voluptatem
+                        voluptaslorem ipsum dolor sit amet consectetur
+                        adipisicing elit. Quisquam quod, voluptatum, quae
+                        voluptatem voluptaslorem ipsum dolor sit amet
+                        consectetur adipisicing elit. Quisquam quod, voluptatum,
+                        quae voluptatem voluptasloremquae voluptatem
+                        voluptaslorem ipsum dolor sit amet consectetur
+                        adipisicing elit. Quisquam quod, voluptatum, quae
+                        voluptatem voluptaslorem ipsum dolor sit amet
+                        consectetur adipisicing elit. Quisquam quod, voluptatum,
+                        quae voluptatem voluptasloremquae voluptatem
+                        voluptaslorem ipsum dolor sit amet consectetur
+                        adipisicing elit. Quisquam quod, voluptatum, quae
+                        voluptatem voluptaslorem ipsum dolor sit amet
+                        consectetur adipisicing elit. Quisquam quod, voluptatum,
+                        quae voluptatem voluptasloremquae voluptatem
+                        voluptaslorem ipsum dolor sit amet consectetur
+                        adipisicing elit. Quisquam quod, voluptatum, quae
+                        voluptatem voluptaslorem ipsum dolor sit amet
+                        consectetur adipisicing elit. Quisquam quod, voluptatum,
+                        quae voluptatem voluptasloremquae voluptatem
+                        voluptaslorem ipsum dolor sit amet consectetur
+                        adipisicing elit. Quisquam quod, voluptatum, quae
+                        voluptatem voluptaslorem ipsum dolor sit amet
+                        consectetur adipisicing elit. Quisquam quod, voluptatum,
+                        quae voluptatem voluptaslorem
+                    </div>
+                </Modal>
+                <button
+                    onClick={() => {
+                        setArgs({ ...args, open: true });
+                    }}
+                >
+                    Open modal
+                </button>
+            </div>
+        );
     },
 };
