+++ src/powerhouse/components/modal/modal.tsx
@@ -1,77 +1,41 @@
-import {
-    Modal as AriaModal,
-    Dialog,
-    DialogProps,
-    ModalOverlay,
-    ModalOverlayProps,
-} from 'react-aria-components';
+import { Content, Overlay, Portal, Root } from '@radix-ui/react-dialog';
+import { ComponentPropsWithoutRef } from 'react';
 import { twMerge } from 'tailwind-merge';
 
-export interface ModalProps
-    extends ModalOverlayProps,
-        React.RefAttributes<HTMLDivElement> {
-    open?: boolean;
-    onClose?: () => void;
+type Props = {
     children?: React.ReactNode;
-    modalProps?: ModalOverlayProps & React.RefAttributes<HTMLDivElement>;
-    dialogProps?: DialogProps & React.RefAttributes<HTMLElement>;
-}
-
-export const Modal: React.FC<ModalProps> = props => {
-    const {
-        onClose,
-        children,
-        open = false,
-        modalProps = {},
-        dialogProps = {},
-        className: modalOverlayClassName,
-        ...modalOverlayProps
-    } = props;
-
-    const { className: ariaModalClassName, ...ariaModalProps } = modalProps;
-    const { className: ariaDialogClassName, ...ariaDialogProps } = dialogProps;
-
-    return (
-        <ModalOverlay
-            className={({ isEntering, isExiting }) =>
-                twMerge(
-                    'bg-overlay fixed inset-0 grid place-items-center',
-                    typeof modalOverlayClassName === 'string' &&
-                        modalOverlayClassName,
-                    isEntering && 'duration-300 ease-out animate-in fade-in',
-                    isExiting && 'duration-200 ease-in animate-out fade-out',
-                )
-            }
-            isDismissable
-            isOpen={open}
-            onOpenChange={onClose}
-            {...modalOverlayProps}
-        >
-            <AriaModal
-                className={({ isEntering, isExiting }) =>
-                    twMerge(
-                        'relative flex justify-center',
-                        typeof ariaModalClassName === 'string' &&
-                            ariaModalClassName,
-                        isEntering &&
-                            'duration-300 ease-out animate-in zoom-in-95',
-                        isExiting &&
-                            'duration-200 ease-in animate-out zoom-out-95',
-                    )
-                }
-                {...ariaModalProps}
-            >
-                <Dialog
-                    className={twMerge(
-                        'bg-white outline-none',
-                        typeof ariaDialogClassName === 'string' &&
-                            ariaDialogClassName,
-                    )}
-                    {...ariaDialogProps}
-                >
-                    {children}
-                </Dialog>
-            </AriaModal>
-        </ModalOverlay>
-    );
+    open?: boolean;
+    onOpenChange?: (open: boolean) => void;
+    overlayProps?: ComponentPropsWithoutRef<typeof Overlay>;
+    contentProps?: ComponentPropsWithoutRef<typeof Content>;
 };
+export function Modal(props: Props) {
+    return (
+        <Root
+            open={props.open}
+            defaultOpen={props.open}
+            onOpenChange={props.onOpenChange}
+        >
+            <Portal>
+                <Overlay
+                    {...props.overlayProps}
+                    className={twMerge(
+                        'fixed inset-0 grid place-items-center overflow-y-auto bg-slate-900/50 data-[state=closed]:animate-fade-out data-[state=open]:animate-fade-in',
+                        props.overlayProps?.className,
+                    )}
+                >
+                    <Content
+                        {...props}
+                        {...props.contentProps}
+                        className={twMerge(
+                            'bg-white data-[state=closed]:animate-zoom-out data-[state=open]:animate-zoom-in',
+                            props.contentProps?.className,
+                        )}
+                    >
+                        {props.children}
+                    </Content>
+                </Overlay>
+            </Portal>
+        </Root>
+    );
+}
