--- src/powerhouse/components/new-modal/index.ts
@@ -1,1 +1,0 @@
-export * from './modal';
